More cleanup
Former-commit-id: 7c826cad3e78b1549b8100804103bf57eaa491bb

+++ WorldsEngine/Render/Render.hpp
@@ -151,8 +151,8 @@ namespace worlds {
         bool active = false;
         float resScale = 1.0f;
 
-        uint32_t actualWidth() { return width * resScale; }
-        uint32_t actualHeight() { return height * resScale; }
+        uint32_t actualWidth() { return (uint32_t)(width * resScale); }
+        uint32_t actualHeight() { return (uint32_t)(height * resScale); }
 
         virtual void drawNow(entt::registry& world) = 0;
         virtual void requestPick(int x, int y) = 0;
